requirements.txt
langgraph==0.2.70
langchain==0.3.18
langchain-chroma==0.2.1
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.3
langchain-teddynote==0.3.42
langchain-upstage==0.6.0
Django==5.1.6
pydantic==2.10.6
python-dotenv==1.0.1
dataclasses==0.6
uuid==1.30
DateTime==5.5
typing==3.7.4.3
